created Company class created CompanyRepository interface
added a pk/fk relationship between user and company.
	edited User and UserRepository for this purpose.

added a user in the WebBasedEvaluationApplication class. I had to create
the user only once and then comment out the code for it. This is the
initial user we can use to login. email: admin@gmail.com and pw:
password

I also found that in order to edit/create/drop tables you must change
the spring.jpa.hibernate.ddl-auto property.