Fixed several problems
added the adm_eval role to the role sheet. This role can be used to
upload the eval templates.

took away the ability for anyone to view/edit the super user account.

made changes to ensure that no one can elevate their own privilege.

changed the super user to not have avccess to eval forms/groups.